fix: network inspector - disable native implementation on android 0.83.0 due to issues (#1816)
### Description

This PR introduces additional check for React Native `0.83.0` running on
an Android device, which helps to disable native network inspector in
this case.

When using native network inspector implementation with Android device
on version `0.83.0`, the `Size` of request made to an `https` endpoint
is always `-1`. This behaviour is caused by React Native DevTools bug
and is not our fault (it is in fact also apparent in the devtools window
itself), but given that our implementation works in this case, we should
switch to it to provide proper user experience when we can.


<img width="1190" height="356" alt="image"
src="https://github.com/user-attachments/assets/12bbf342-fb2f-4b9f-be2e-2f0618efd54d"
/>

This change required moving the compatibility check from
`ApplicationContext` to `ApplicationSession`, as we do not have
information about used device in the LaunchConfig (it is tailored for
both platforms, the platform information itself comes at later
initialisation stage).

### How Has This Been Tested: 

Verified that, on version `0.83.0` native network inspector is used on
iOS and disabled on android, switched to other version (`0.82.0`), then
back to `0.83.0` to check whether the version updates correctly.

Changed `package.json` of react-native to state that version is `0.83.2`
and checked that in this case, android properly uses the native
implementation.

Checked that explicitly toggling the option
`useNativeInspectorImplementation` in launch configuration works as
expected.

### How Has This Change Been Documented:

Not Applicable.