fix: patch the element inspector for React `19.0.1` in React Native 78 and 79 (#1815)
### Description

Fix the element inspector being broken when using React `19.0.1` due to
adfefc0 patch.

In adfefc0, `babel_transofrmer` was updated to only replace the
`JSXRuntime` in RN78 and RN79 when React version is `19.0.0`, making it
impossible for element inspector to work due to reasons explain in the
commit message.

This PR reverts those changes, allowing for the JSX runtime to be
replaced, by modifying the version check for `19.0.0` in our modified
bundle for `ReactNativeRenderer-dev` (for information why it is modified
in the first place, see
https://github.com/software-mansion/radon-ide/commit/ee393ecf4e57b3c76252c266b74df4dabbf5afcc)

The change was made to the version check, which checked whether the
version of React is `19.0.0`

```ts
// ...
if ("19.0.0" !== isomorphicReactPackageVersion) {
  // error throw
}
``` 

EDIT [Now handling versions 19.0.X]
In order to handle the situation, in which someone bumps their React
version to `19.0.1`, `19.0.2`,`19.0.3` or any `19.0.X`, we modified our
bundled renderer.

```ts
// ...
if (!isomorphicReactPackageVersion.startsWith("19.0")) {
  // error throw
}
``` 

This is possible, due to the fact that the only changes introduced to
`react` package between `19.0.0` and the patches are version bumps in
some of the files (see https://npmdiff.dev/react/19.0.0/19.0.3/), so
extending the check within the `ReactNativeRenderer` to also allow for
`19.0.X` is safe and should not introduce any problems. This approach
allows us to support the element inspector, unlike the previous fix.

### How Has This Been Tested: 

- Verified that existing projects using React Native 78 and 79 with
React `19.0.0` run correctly
- Replaced the version of React in those projects to be `19.0.1` and
`19.0.3`, ran
`npm install`, verified the app and element inspector work correctly
- Created fresh React Native 78 and 79 using react-native-cli to confirm
that fresh apps with `19.0.0` remain working correctly

### How Has This Change Been Documented:

Not applicable.