feat: Improve telemetry for editor type and license type (#1828)
The editor type is currently embedded inside the event name. That
complicates data analysis.

This PR fixes this issue, and adds data on the active license type
(`Free`, `Pro`, `Inactive`, and others).

**Note:** These tests are, and always will be flaky. This is
unavoidable. The goal of these tests isn't to get `100%` success rate,
the goal is to improve the current success rate from `~0%` to `~70%`.

### How Has This Been Tested:

- Checked input properties at the time of `sendTelemetryEvent` call
- All properties match the expected value

### How Has This Change Been Documented:

Not applicable