fix: loading metro config with new (newer then 54.0.20) Expo CLI (#1834)
Tis PR solves the a problem with loading metro configuration in Radon,
that was introduced with [this
change](https://github.com/expo/expo/pull/41142) in expo CLI.

Previously Expo CLI has relied on `loadConfig`, that we have overridden
to adapt the config. After the change mentioned above expo decided to
use `resolveConfig` instead, so we temporally override it as well.

To avoid overriding even more expos internal functionality. We check the
expo CLI version to determine if `resolveConfig` should be overridden.
An alternative solution would be to override both `resolveConfig` and
`getDefaultConfig` and modify the logic of `loadConfig` to merge the
changes correctly after being given already modified `default`, but this
option would potentially affect how the old setups behave, which we want
to avoid.

Most of the changes made in this PR should be removed once expo
implements more stable solution for providing custom `metro-configs`.

### How Has This Been Tested: 

- run expo app with expo version 54.0.30 (this matches the CLI version
54.0.20) - radon should work after this fix
- run expo app with expo version 54.0.29 (this matches the CLI version
54.0.19) - radon should work after this fix and before
- run expo app with expo version 53 - radon should work after this fix
and before

NOTE: If you test multiple versions by reinstalling expo in the same
project you need to use "restart metro" option from radons reload menu

### How Has This Change Been Documented:

internal